# api/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "api",
    "description": "",
    "license": {
      "name": ""
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/galaxies/{galaxy_id}/planets": {
      "get": {
        "tags": [
          "routes::planet"
        ],
        "operationId": "get_all_planets",
        "parameters": [
          {
            "name": "galaxy_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/PlanetsList"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizeResponse"
          },
          "404": {
            "$ref": "#/components/responses/NotFoundResponse"
          },
          "500": {
            "$ref": "#/components/responses/InternalErrorResponse"
          }
        }
      },
      "post": {
        "tags": [
          "routes::planet"
        ],
        "operationId": "create_planet",
        "parameters": [
          {
            "name": "galaxy_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "description": "data for creating the planet",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePlanetData"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/PlanetCreated"
          },
          "400": {
            "$ref": "#/components/responses/ValidationResponse"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizeResponse"
          },
          "404": {
            "$ref": "#/components/responses/NotFoundResponse"
          },
          "409": {
            "$ref": "#/components/responses/AlreadyExistsResponse"
          },
          "500": {
            "$ref": "#/components/responses/InternalErrorResponse"
          }
        }
      }
    },
    "/galaxies/{galaxy_id}/planets/{planet_id}": {
      "get": {
        "tags": [
          "routes::planet"
        ],
        "operationId": "get_planet",
        "parameters": [
          {
            "name": "galaxy_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "planet_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/SpecificPlanet"
          },
          "400": {
            "$ref": "#/components/responses/ValidationResponse"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizeResponse"
          },
          "404": {
            "$ref": "#/components/responses/NotFoundResponse"
          },
          "409": {
            "$ref": "#/components/responses/AlreadyExistsResponse"
          },
          "500": {
            "$ref": "#/components/responses/InternalErrorResponse"
          }
        }
      },
      "put": {
        "tags": [
          "routes::planet"
        ],
        "operationId": "update_planet",
        "parameters": [
          {
            "name": "galaxy_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "planet_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "description": "data for updating the planet",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePlanetData"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/PlanetUpdated"
          },
          "400": {
            "$ref": "#/components/responses/ValidationResponse"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizeResponse"
          },
          "404": {
            "$ref": "#/components/responses/NotFoundResponse"
          },
          "409": {
            "$ref": "#/components/responses/AlreadyExistsResponse"
          },
          "500": {
            "$ref": "#/components/responses/InternalErrorResponse"
          }
        }
      },
      "delete": {
        "tags": [
          "routes::planet"
        ],
        "operationId": "delete_planet",
        "parameters": [
          {
            "name": "galaxy_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "planet_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/PlanetDeleted"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizeResponse"
          },
          "404": {
            "$ref": "#/components/responses/NotFoundResponse"
          },
          "500": {
            "$ref": "#/components/responses/InternalErrorResponse"
          }
        }
      }
    },
    "/galaxies/{galaxy_id}/stars": {
      "get": {
        "tags": [
          "routes::star"
        ],
        "operationId": "get_all_stars",
        "parameters": [
          {
            "name": "galaxy_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/StarsList"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizeResponse"
          },
          "404": {
            "$ref": "#/components/responses/NotFoundResponse"
          },
          "500": {
            "$ref": "#/components/responses/InternalErrorResponse"
          }
        }
      },
      "post": {
        "tags": [
          "routes::star"
        ],
        "operationId": "create_star",
        "parameters": [
          {
            "name": "galaxy_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "description": "data for creating the star",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateStarData"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/StarCreated"
          },
          "400": {
            "$ref": "#/components/responses/ValidationResponse"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizeResponse"
          },
          "404": {
            "$ref": "#/components/responses/NotFoundResponse"
          },
          "409": {
            "$ref": "#/components/responses/AlreadyExistsResponse"
          },
          "500": {
            "$ref": "#/components/responses/InternalErrorResponse"
          }
        }
      }
    },
    "/galaxies/{galaxy_id}/stars/{star_id}": {
      "get": {
        "tags": [
          "routes::star"
        ],
        "operationId": "get_star",
        "parameters": [
          {
            "name": "galaxy_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "star_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/SpecificStar"
          },
          "400": {
            "$ref": "#/components/responses/ValidationResponse"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizeResponse"
          },
          "404": {
            "$ref": "#/components/responses/NotFoundResponse"
          },
          "409": {
            "$ref": "#/components/responses/AlreadyExistsResponse"
          },
          "500": {
            "$ref": "#/components/responses/InternalErrorResponse"
          }
        }
      },
      "put": {
        "tags": [
          "routes::star"
        ],
        "operationId": "update_star",
        "parameters": [
          {
            "name": "galaxy_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "star_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "description": "data for updating the star",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateStarData"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/StarUpdated"
          },
          "400": {
            "$ref": "#/components/responses/ValidationResponse"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizeResponse"
          },
          "404": {
            "$ref": "#/components/responses/NotFoundResponse"
          },
          "409": {
            "$ref": "#/components/responses/AlreadyExistsResponse"
          },
          "500": {
            "$ref": "#/components/responses/InternalErrorResponse"
          }
        }
      },
      "delete": {
        "tags": [
          "routes::star"
        ],
        "operationId": "delete_star",
        "parameters": [
          {
            "name": "galaxy_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "star_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/StarDeleted"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizeResponse"
          },
          "404": {
            "$ref": "#/components/responses/NotFoundResponse"
          },
          "500": {
            "$ref": "#/components/responses/InternalErrorResponse"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ConnectPlanetToStar": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          }
        }
      },
...
      "CreatePlanetData": {
        "type": "object",
        "required": [
          "name",
          "capacity",
          "star"
        ],
        "properties": {
          "capacity": {
            "type": "integer",
            "format": "int32",
            "maximum": 2147483647,
            "minimum": 0
          },
          "name": {
            "type": "string",
            "minLength": 1
          },
          "star": {
            "$ref": "#/components/schemas/ConnectPlanetToStar"
          }
        }
      },
      "CreateStarData": {
        "type": "object",
        "required": [
          "name",
          "nebula"
        ],
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1
          },
          "nebula": {
            "type": "string",
            "format": "uri"
          }
        }
      },
...
      "Planet": {
        "type": "object",
        "required": [
          "id",
          "name",
          "capacity",
          "galaxy_id"
        ],
        "properties": {
          "capacity": {
            "type": "integer",
            "format": "int32",
            "maximum": 2147483647,
            "minimum": 0
          },
          "galaxy_id": {
            "type": "string",
            "format": "uuid"
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string",
            "minLength": 1
          },
          "star_id": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          }
        }
      },
      "Star": {
        "type": "object",
        "required": [
          "id",
          "name",
          "nebula",
          "galaxy_id"
        ],
        "properties": {
          "galaxy_id": {
            "type": "string",
            "format": "uuid"
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string",
            "minLength": 1
          },
          "nebula": {
            "type": "string",
            "format": "uri"
          }
        }
      },
...
      "UpdatePlanetData": {
        "type": "object",
        "properties": {
          "capacity": {
            "type": "integer",
            "format": "int32",
            "nullable": true,
            "maximum": 2147483647,
            "minimum": 0
          },
          "name": {
            "type": "string",
            "nullable": true,
            "minLength": 1
          },
          "star": {
            "allOf": [
              {
                "$ref": "#/components/schemas/ConnectPlanetToStar"
              }
            ],
            "nullable": true
          }
        }
      },
      "UpdateStarData": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "nullable": true,
            "minLength": 1
          },
          "nebula": {
            "type": "string",
            "format": "uri",
            "nullable": true
          }
        }
      },
...
    },
    "responses": {
      "AlreadyExistsResponse": {
        "description": "The resource already exists",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "required": [
                "status_code",
                "message"
              ],
              "properties": {
                "message": {
                  "type": "string"
                },
                "status_code": {
                  "type": "integer",
                  "format": "int32",
                  "minimum": 0
                }
              }
            }
          }
        }
      },
...
      "InternalErrorResponse": {
        "description": "An internal error occurred",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "required": [
                "status_code",
                "message"
              ],
              "properties": {
                "message": {
                  "type": "string"
                },
                "status_code": {
                  "type": "integer",
                  "format": "int32",
                  "minimum": 0
                }
              }
            }
          }
        }
      },
      "NotFoundResponse": {
        "description": "Requested resources not found",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "required": [
                "status_code",
                "message"
              ],
              "properties": {
                "message": {
                  "type": "string"
                },
                "status_code": {
                  "type": "integer",
                  "format": "int32",
                  "minimum": 0
                }
              }
            }
          }
        }
      },
      "PlanetCreated": {
        "description": "planet successfully created",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Planet"
            }
          }
        }
      },
      "PlanetDeleted": {
        "description": "planet successfully deleted",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Planet"
            }
          }
        }
      },
      "PlanetUpdated": {
        "description": "planet successfully updated",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Planet"
            }
          }
        }
      },
      "PlanetsList": {
        "description": "all planets in the galaxy",
        "content": {
          "application/json": {
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/components/schemas/Planet"
              }
            }
          }
        }
      },
...
      "SpecificPlanet": {
        "description": "specific planet",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Planet"
            }
          }
        }
      },
      "SpecificStar": {
        "description": "specific star",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Star"
            }
          }
        }
      },
      "StarCreated": {
        "description": "star successfully created",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Star"
            }
          }
        }
      },
      "StarDeleted": {
        "description": "star successfully deleted",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Star"
            }
          }
        }
      },
      "StarUpdated": {
        "description": "star successfully updated",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Star"
            }
          }
        }
      },
      "StarsList": {
        "description": "all stars in the galaxy",
        "content": {
          "application/json": {
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/components/schemas/Star"
              }
            }
          }
        }
      },
      "UnauthorizeResponse": {
        "description": "User not authorized",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "required": [
                "status_code",
                "message"
              ],
              "properties": {
                "message": {
                  "type": "string"
                },
                "status_code": {
                  "type": "integer",
                  "format": "int32",
                  "minimum": 0
                }
              }
            }
          }
        }
      },
...
      "ValidationResponse": {
        "description": "The body of the request contains incorrect data",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "required": [
                "status_code",
                "message"
              ],
              "properties": {
                "message": {
                  "type": "string"
                },
                "status_code": {
                  "type": "integer",
                  "format": "int32",
                  "minimum": 0
                }
              }
            }
          }
        }
      }
    },
    "securitySchemes": {
      "session_token": {
        "type": "apiKey",
        "in": "cookie",
        "name": "session"
      }
    }
  },
  "security": [
    {
      "session_token": []
    }
  ]
}